requirements.txt
fastapi~=0.115.4
uvicorn[standard]~=0.30.6
httpx~=0.27.2
python-dotenv~=1.0.1
pytz~=2024.2
sqlalchemy~=2.0.36
alembic~=1.13.2
psycopg2-binary~=2.9.9
pydantic~=2.9.2
pytest~=8.3.3
